requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
python-dotenv>=1.0.0

